+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## 1.3.0: 2019-08-08
+
+- Adds an email confirmation option post-submission. Improves error handling with message label. Arranges fields in form field widgets. Prevents `apostrophe-anchors` from adding fields to each form input widget.
+
+## 1.2.1: 2019-07-22
+
+- Added error handling for misconfigured email submission sending.
+
 ## 1.2.0: 2019-07-22
 
 - Adds the conditional field group widget, availble to be triggered by boolean, select, or radio fields.